+++ Product/MySerial.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "WString.h"
 
 #define SERIAL_ENABLED
 // It's probably safe to #define SERIAL_ENABLED in the product build,
@@ -8,9 +9,13 @@
 #ifdef SERIAL_ENABLED
 void serialInit(bool inputAllowed);
 void serialPrintf(const char* __fmt, ...);
+void serialPrint(const __FlashStringHelper* string);
+void serialPrintln(const __FlashStringHelper* string);
 char* renderLongLong(long long num);
 #else
 #define serialInit()
 #define serialPrintf(...)
 #define renderLongLong(...)
+#define serialPrint(...)
+#define serialPrintln(...)
 #endif
